Rename the package from '...model.person' to '...model.task'
