+++ src/main/java/seedu/address/logic/commands/SelectCommand.java
@@ -3,8 +3,8 @@ package seedu.address.logic.commands;
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.events.ui.JumpToListRequestEvent;
+import seedu.address.model.task.ReadOnlyTask;
 import seedu.address.commons.core.UnmodifiableObservableList;
-import seedu.address.model.person.ReadOnlyTask;
 
 /**
  * Selects a person identified using it's last displayed index from the address book.
