+++ src/main/java/seedu/address/model/ReadOnlyTaskList.java
@@ -1,10 +1,10 @@
 package seedu.address.model;
 
 
-import seedu.address.model.person.ReadOnlyTask;
-import seedu.address.model.person.UniqueTaskList;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
+import seedu.address.model.task.ReadOnlyTask;
+import seedu.address.model.task.UniqueTaskList;
 
 import java.util.List;
