+++ src/main/java/seedu/address/model/task/UniqueTaskList.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.task;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
