+++ src/test/java/guitests/SelectCommandTest.java
@@ -1,7 +1,8 @@
 package guitests;
 
 import org.junit.Test;
-import seedu.address.model.person.ReadOnlyTask;
+
+import seedu.address.model.task.ReadOnlyTask;
 
 import static org.junit.Assert.assertEquals;
