+++ src/test/java/seedu/address/storage/XmlAddressBookStorageTest.java
@@ -8,8 +8,8 @@ import org.junit.rules.TemporaryFolder;
 import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.model.TaskList;
+import seedu.address.model.task.Task;
 import seedu.address.model.ReadOnlyTaskList;
-import seedu.address.model.person.Task;
 import seedu.address.testutil.TypicalTestTasks;
 
 import java.io.IOException;
