+++ src/test/java/seedu/address/testutil/TaskBuilder.java
@@ -2,7 +2,7 @@ package seedu.address.testutil;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.person.*;
+import seedu.address.model.task.*;
 
 /**
  *
