+++ src/test/java/seedu/address/testutil/TestTask.java
@@ -1,7 +1,7 @@
 package seedu.address.testutil;
 
 import seedu.address.model.tag.UniqueTagList;
-import seedu.address.model.person.*;
+import seedu.address.model.task.*;
 
 /**
  * A mutable task object. For testing only.
