+++ src/test/java/seedu/address/testutil/TypicalTestTasks.java
@@ -2,7 +2,7 @@ package seedu.address.testutil;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.TaskList;
-import seedu.address.model.person.*;
+import seedu.address.model.task.*;
 
 /**
  *
